+++ lib/assets/components/buttons/darkButton.dart
@@ -1,8 +1,39 @@
 import 'package:app/assets/StyleFrom.dart';
 import 'package:app/assets/components/customTile.dart';
 import 'package:flutter/material.dart';
-import 'package:app/assets/components/buttons/buttons_styles.dart';
 
+final double defaultBorderRadius = 8.0;
+final double defaultPadding = 14.0;
+
+enum DarkButtonSize { Small, Medium, Large }
+enum DarkButtonStyle { Primary, Secondary, Outline }
+
+Map darkButtonStyleStyles = {
+  DarkButtonSize.Small: {
+    'style': DarkButtonSize.Small,
+    'height': 32.0,
+    'borderRadius': 8.0,
+    'fontSize': 14.0,
+    'hPadding': 8,
+    'vPadding': 8,
+  },
+  DarkButtonSize.Medium: {
+    'style': DarkButtonSize.Medium,
+    'height': 40.0,
+    'borderRadius': 8.0,
+    'fontSize': 17.0,
+    'hPadding': 16.0,
+    'vPadding': 8.0,
+  },
+  DarkButtonSize.Large: {
+    'style': DarkButtonSize.Large,
+    'height': 48.0,
+    'borderRadius': 8.0,
+    'fontSize': 17.0,
+    'hPadding': 16.0,
+    'vPadding': 8.0,
+  },
+};
 
 // ignore: must_be_immutable
 class DarkButton extends StatelessWidget {
@@ -10,12 +41,12 @@ class DarkButton extends StatelessWidget {
   final String text;
   final IconData rightIcon;
   final IconData leftIcon;
-  final ButtonSize size;
+  final DarkButtonSize size;
   final double fontSize;
   final DarkButtonStyle style;
 
   DarkButton(this.text,
-      {this.size = ButtonSize.Medium,
+      {this.size = DarkButtonSize.Medium,
       @required this.onPressed,
       this.rightIcon,
       this.leftIcon,
@@ -55,7 +86,7 @@ class SecondaryButton extends StatelessWidget {
   final VoidCallback onPressed;
   final String text;
   final IconData rightIcon;
-  final ButtonSize size;
+  final DarkButtonSize size;
   final double fontSize;
 
   SecondaryButton(
@@ -105,7 +136,7 @@ class PrimaryButton extends StatelessWidget {
   final VoidCallback onPressed;
   final String text;
   final IconData rightIcon;
-  final ButtonSize size;
+  final DarkButtonSize size;
   final double fontSize;
 
   PrimaryButton(
@@ -170,7 +201,7 @@ class PrimaryButton extends StatelessWidget {
 class OutlineButton extends StatelessWidget {
   final VoidCallback onPressed;
   final String text;
-  final ButtonSize size;
+  final DarkButtonSize size;
   final double fontSize;
 
   OutlineButton(
@@ -220,106 +251,4 @@ class OutlineButton extends StatelessWidget {
   }
 }
 
-class CustomWidthButton extends StatelessWidget {
-  final VoidCallback onPressed;
-  final String text;
-  final IconData rightIcon;
-  final ButtonSize size;
-  final double fontSize;
-  final double buttonWidthProportion;
-  final Color backgroundColor;
-  final Color textColor;
 
-  CustomWidthButton(this.text,
-      {@required this.onPressed,
-      this.rightIcon,
-      @required this.size,
-      @required this.fontSize,
-      @required this.buttonWidthProportion,
-      this.backgroundColor,
-      this.textColor});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-        width: MediaQuery.of(context).size.width * buttonWidthProportion,
-        child: MaterialButton(
-          height: darkButtonStyleStyles[size]['height'],
-          minWidth: buttonWidthProportion,
-          elevation: 0,
-          color: this.backgroundColor ?? Colors.white,
-          disabledColor: colorFrom(
-            Theme.of(context).primaryColor,
-            opacity: 0.5,
-          ),
-          onPressed: onPressed,
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(
-                darkButtonStyleStyles[size]['borderRadius']),
-          ),
-          child: Padding(
-            padding: EdgeInsets.symmetric(
-              vertical: darkButtonStyleStyles[size]['vPadding'],
-              horizontal: darkButtonStyleStyles[size]['hPadding'],
-            ),
-            child: Text(
-              text,
-              style: textStyleFrom(
-                Theme.of(context).primaryTextTheme.button,
-                color: textColor ?? Colors.white,
-                fontSize: fontSize ?? darkButtonStyleStyles[size]['fontSize'],
-              ),
-            ),
-          ),
-        ));
-  }
-}
-
-class CustomIconButton extends StatelessWidget {
-  final VoidCallback onPressed;
-  final IconData icon;
-  final ButtonSize size;
-  final Color backgroundColor;
-  final Color iconColor;
-  final bool isCircularButton;
-
-  CustomIconButton({
-    @required this.onPressed,
-    @required this.icon,
-    this.iconColor,
-    @required this.size,
-    @required this.isCircularButton,
-    this.backgroundColor,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialButton(
-      height: darkButtonStyleStyles[size]['height'],
-      elevation: 0,
-      color: this.backgroundColor ?? Colors.white,
-      disabledColor: colorFrom(
-        Theme.of(context).primaryColor,
-        opacity: 0.5,
-      ),
-      onPressed: onPressed,
-      shape: isCircularButton
-          ? CircleBorder()
-          : RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(
-                  darkButtonStyleStyles[size]['borderRadius']),
-            ),
-      padding: isCircularButton
-          ? EdgeInsets.all(0)
-          : EdgeInsets.symmetric(horizontal: 60.0, vertical: 15.0),
-      child: Padding(
-        padding: isCircularButton ? const EdgeInsets.all(5) : EdgeInsets.all(0),
-        child: Icon(
-          icon,
-          size: isCircularButton ? 15 : 25,
-          color: iconColor ?? Theme.of(context).primaryColor,
-        ),
-      ),
-    );
-  }
-}
