+++ lib/managers/dialog_manager.dart
@@ -4,7 +4,7 @@ import 'package:app/locator.dart';
 import 'package:app/services/dialog_service.dart';
 import 'package:app/services/navigation_service.dart';
 
-import 'package:app/assets/components/darkButton.dart';
+import 'package:app/assets/components/buttons/darkButton.dart';
 import 'package:app/assets/StyleFrom.dart';
 
 class DialogManager extends StatefulWidget {
