+++ lib/pages/other/NotificationPage.dart
@@ -1,4 +1,4 @@
-import 'package:app/assets/components/darkButton.dart';
+import 'package:app/assets/components/buttons/darkButton.dart';
 import 'package:flutter/material.dart';
 
 import 'package:app/assets/StyleFrom.dart';
